Replace Array::ArrayIteratorImpl_ with BufferIterator
