+++ tests/cpp/meson.build
@@ -24,7 +24,6 @@ tests_all = executable(
         'nodes/test_quadratic_model.cpp',
         'test_array.cpp',
         'test_graph.cpp',
-        'test_iterators.cpp',
         'test_simplex.cpp',
         'test_utils.cpp',
     ],
